+++ Xyperico.Agres/Xyperico.Agres/IEventStore.cs
@@ -7,6 +7,6 @@ namespace Xyperico.Agres
   public interface IEventStore
   {
     EventStream Load(IIdentity id);
-    void Append(IIdentity id, int expectedVersion, IEnumerable<IEvent> events);
+    void Append(IIdentity id, long expectedVersion, IEnumerable<IEvent> events);
   }
 }
